refactor =>
change resolveExternalValue => createExternalValue
