+++ packages/delegate/src/index.ts
@@ -6,8 +6,8 @@ export * from './delegateToSchema';
 export * from './delegationBindings';
 export * from './externalObjects';
 export * from './externalValues';
+export * from './mergeDataAndErrors';
 export * from './getMergedParent';
-export * from './resolveExternalValue';
 export * from './subschemaConfig';
 export * from './transforms';
 export * from './types';
